Changed to a proper maven layout
