+++ .gitignore
@@ -1,0 +1,5 @@
+/bin/
+.idea
+target
+*.iml
+dependency-reduced-pom.xml
