--- CrazyAdvancementAPI/pom.xml
@@ -1,34 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>eu.endercentral.crazy_advancements</groupId>
-  <artifactId>CrazyAdvancementsAPI</artifactId>
-  <version>1.13.6</version>
-  <packaging>jar</packaging>
-
-  <name>CrazyAdvancementsAPI</name>
-  <url>https://crazyadvancements.endercentral.eu</url>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <dependencies>
-    <dependency>
-           <groupId>org.spigotmc</groupId>
-           <artifactId>spigot-api</artifactId>
-           <version>1.15-R0.1-SNAPSHOT</version>
-           <scope>provided</scope>
-    </dependency>
-  </dependencies>
-  
-  <repositories>
-	 <repository>
-        <id>spigot-repo</id>
-        <url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
-    </repository>
-  </repositories>
-  
-  
-</project>
