+++ src/main/java/eu/endercentral/crazy_advancements/AdvancementReward.java
@@ -1,0 +1,9 @@
+package eu.endercentral.crazy_advancements;
+
+import org.bukkit.entity.Player;
+
+public abstract class AdvancementReward {
+	
+	public abstract void onGrant(Player player);
+	
+}
